requirements.txt
spotipy
notion-client
python-dotenv